+++ apps/journeys-admin/src/components/Team/TeamUpdateDialog/TeamUpdateDialog.spec.tsx
@@ -13,7 +13,7 @@ import {
 import { TEAM_UPDATE } from './TeamUpdateDialog'
 
 import { TeamUpdateDialog } from '.'
-import '../../../../test/i18n'
+import '../../../../../../libs/journeys/ui/test/i18n'
 
 describe('TeamUpdateDialog', () => {
   const getTeamsMock: MockedResponse<GetLastActiveTeamIdAndTeams> = {
