+++ apps/journeys-admin/src/components/TemplateList/TrashedTemplateList/TrashedTemplateList.stories.tsx
@@ -18,7 +18,7 @@ import {
 
 import { TrashedTemplateList } from '.'
 
-import '../../../../test/i18n'
+import '../../../../../../libs/journeys/ui/test/i18n'
 
 const TrashedTemplateListStory: Meta<typeof TrashedTemplateList> = {
   component: TrashedTemplateList,
