+++ apps/journeys-admin/src/components/VisitorInfo/VisitorJourneysList/EventsCard/TimelineEvent/TimelineEvent.spec.tsx
@@ -19,7 +19,7 @@ import {
 } from '../../utils/data'
 
 import { TimelineEvent } from '.'
-import '../../../../../../test/i18n'
+import '../../../../../../../../libs/journeys/ui/test/i18n'
 
 describe('TimelineEvent', () => {
   it('shows journeyViewEvent', () => {
